step 5
Made a function which prints out the results of each game. I also made some edits to the main code, removing Meagan's function since she had already moved it to its own file. Her file was also calling the function, causing it to repeat twice when you played the game, so I got rid of that.